==> py data clean/07_retail_store_visits.ipynb <==
import requests
import pandas as pd
import sys
from io import BytesIO
import gzip
import io
import os
from global_variables import *
import sqlite3
from urllib.parse import urlparse

poi_sod = pd.read_csv("C:/Users/dratnadiwakara2/Documents/OneDrive - Louisiana State University/Raw Data/Dewey/poi_sod_4.csv.gz", compression='gzip')

poi = pd.read_csv("C:/Users/dratnadiwakara2/Documents/OneDrive - Louisiana State University/Raw Data/Dewey/all_us_poi_with_store_id.csv.gz", compression='gzip',usecols=['PLACEKEY','POSTAL_CODE','STREET_ADDRESS','STORE_ID'])

poi_sod = pd.merge(poi, poi_sod, on='PLACEKEY', how='inner')

del poi

poi_sod['match_address'] = poi_sod['STREET_ADDRESS'].str.lower()+' '+poi_sod['POSTAL_CODE'].astype(int).astype(str)
unique_match_addresses = poi_sod['match_address'].unique()

poi_sod['POSTAL_CODE'] = poi_sod['POSTAL_CODE'].astype(int)

poi_sod.head()

results = requests.get(url=Retail_Store_Visitors_API_PATH,
                           params={'page': 1,
                                   'partition_key_after': '1900-01-01', # set date value here
                               'partition_key_before': '2099-01-01'},
                           headers={'X-API-KEY': API_KEY,
                                    'accept': 'application/json'
                                   })
response_json = results.json()

data_path = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_banks.csv'
for link_data in response_json['download_links'][195:]:
    print(f"Downloading file {link_data['file_name']}...")
    data = requests.get(link_data['link'])
    gzip_stream = io.BytesIO(data.content)
        
    df = pd.read_csv(gzip_stream, compression='gzip', low_memory=False)
    df['match_address'] = df['STREET_ADDRESS'].str.lower()+' '+df['ZIP_CODE'].astype(str)
    
    filtered_df = df[df['match_address'].isin(unique_match_addresses)]
    
    keywords = ['bank', 'finan', 'trust', 'capital']
    mask = filtered_df['BRAND'].str.lower().str.contains('|'.join(keywords), case=False)
    
    filtered_df = filtered_df[mask]
    
    if os.path.isfile(data_path):
        header_option = False  
    else:
        header_option = True  
    
    filtered_df.to_csv(data_path, mode='a', header=header_option, index=False)     

ft_data = pd.read_csv(data_path, low_memory=False)


ft_data = ft_data.reset_index()
with gzip.open(data_path+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    ft_data.to_csv(gzipped_file, index=False)




Identify individual components

from tqdm import tqdm
import json

ft_data = pd.read_csv("C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_banks.csv.gz", compression='gzip')

ft_data.info()

VISITOR_HOUSEHOLD_INCOME_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_VISITOR_HOUSEHOLD_INCOME.csv'
VISITOR_EDUCATIONAL_ATTAINMENT_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_VISITOR_EDUCATIONAL_ATTAINMENT.csv'
VISITOR_AGE_RANGE_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_VISITOR_AGE_RANGE.csv'
VISITOR_PSYCHOGRAPHIC_PROFILES_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_VISITOR_PSYCHOGRAPHIC_PROFILES.csv'

with tqdm(total=len(ft_data)) as pbar:
  
     for index in range(0, len(ft_data)):
        t = ft_data.loc[index]

        data_string = t['VISITOR_HOUSEHOLD_INCOME'].replace('},{', '_') .replace("{{","{").replace("}}","}")
        data_string = data_string.replace(",",":").replace("_",",")
        data_dict = json.loads(data_string)
        df_VISITOR_HOUSEHOLD_INCOME = pd.DataFrame.from_dict(data_dict, orient='index', columns=['Value']).reset_index()
        df_VISITOR_HOUSEHOLD_INCOME = df_VISITOR_HOUSEHOLD_INCOME.rename(columns={'index': 'VISITOR_HOUSEHOLD_INCOME', 'Value': 'no_visits'})
        df_VISITOR_HOUSEHOLD_INCOME['index']=t['index']
        df_VISITOR_HOUSEHOLD_INCOME['STORE_ID']=t['STORE_ID']
        
        
        data_string = t['VISITOR_EDUCATIONAL_ATTAINMENT'].replace('},{', '_') .replace("{{","{").replace("}}","}")
        data_string = data_string.replace(",",":").replace("_",",")
        data_dict = json.loads(data_string)
        df_VISITOR_EDUCATIONAL_ATTAINMENT = pd.DataFrame.from_dict(data_dict, orient='index', columns=['Value']).reset_index()
        df_VISITOR_EDUCATIONAL_ATTAINMENT = df_VISITOR_EDUCATIONAL_ATTAINMENT.rename(columns={'index': 'VISITOR_EDUCATIONAL_ATTAINMENT', 'Value': 'no_visits'})
        df_VISITOR_EDUCATIONAL_ATTAINMENT['index']=t['index']
        df_VISITOR_EDUCATIONAL_ATTAINMENT['STORE_ID']=t['STORE_ID']
        
        
        data_string = t['VISITOR_AGE_RANGE'].replace('},{', '_') .replace("{{","{").replace("}}","}")
        data_string = data_string.replace(",",":").replace("_",",")
        data_dict = json.loads(data_string)
        df_VISITOR_AGE_RANGE = pd.DataFrame.from_dict(data_dict, orient='index', columns=['Value']).reset_index()
        df_VISITOR_AGE_RANGE = df_VISITOR_AGE_RANGE.rename(columns={'index': 'VISITOR_AGE_RANGE', 'Value': 'no_visits'})
        df_VISITOR_AGE_RANGE['index']=t['index']
        df_VISITOR_AGE_RANGE['STORE_ID']=t['STORE_ID']
        
                
        data_string = t['VISITOR_PSYCHOGRAPHIC_PROFILES'].replace('},{', '_') .replace("{{","{").replace("}}","}")
        data_string = data_string.replace(",",":").replace("_",",")
        data_dict = json.loads(data_string)
        df_VISITOR_PSYCHOGRAPHIC_PROFILES = pd.DataFrame.from_dict(data_dict, orient='index', columns=['Value']).reset_index()
        df_VISITOR_PSYCHOGRAPHIC_PROFILES = df_VISITOR_PSYCHOGRAPHIC_PROFILES.rename(columns={'index': 'VISITOR_PSYCHOGRAPHIC_PROFILES', 'Value': 'no_visits'})
        df_VISITOR_PSYCHOGRAPHIC_PROFILES['NewColumn'] = df_VISITOR_PSYCHOGRAPHIC_PROFILES['VISITOR_PSYCHOGRAPHIC_PROFILES'].str.extract('(\d+)')
        df_VISITOR_PSYCHOGRAPHIC_PROFILES['NewColumn'] = pd.to_numeric(df_VISITOR_PSYCHOGRAPHIC_PROFILES['NewColumn'])
        df_VISITOR_PSYCHOGRAPHIC_PROFILES = df_VISITOR_PSYCHOGRAPHIC_PROFILES.groupby('NewColumn')['no_visits'].sum().reset_index()
        df_VISITOR_PSYCHOGRAPHIC_PROFILES = df_VISITOR_PSYCHOGRAPHIC_PROFILES.rename(columns={'NewColumn': 'VISITOR_PSYCHOGRAPHIC_PROFILES', 'no_visits': 'no_visits'})
        df_VISITOR_PSYCHOGRAPHIC_PROFILES['index']=t['index']
        df_VISITOR_PSYCHOGRAPHIC_PROFILES['STORE_ID']=t['STORE_ID']
        
        
        if os.path.isfile(VISITOR_HOUSEHOLD_INCOME_file):
            header_option = False  
        else:
            header_option = True
            
        df_VISITOR_HOUSEHOLD_INCOME.to_csv(VISITOR_HOUSEHOLD_INCOME_file, index=False, mode='a', header=header_option,sep="|")
        
        df_VISITOR_EDUCATIONAL_ATTAINMENT.to_csv(VISITOR_EDUCATIONAL_ATTAINMENT_file, index=False, mode='a', header=header_option,sep="|")
        
        df_VISITOR_AGE_RANGE.to_csv(VISITOR_AGE_RANGE_file, index=False, mode='a', header=header_option,sep="|")
         
        df_VISITOR_PSYCHOGRAPHIC_PROFILES.to_csv(VISITOR_PSYCHOGRAPHIC_PROFILES_file, index=False, mode='a', header=header_option,sep="|")
       
        pbar.update(1)

df_VISITOR_HOUSEHOLD_INCOME = pd.read_csv(VISITOR_HOUSEHOLD_INCOME_file)
with gzip.open(VISITOR_HOUSEHOLD_INCOME_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    df_VISITOR_HOUSEHOLD_INCOME.to_csv(gzipped_file, index=False,sep="|")
    
df_VISITOR_EDUCATIONAL_ATTAINMENT = pd.read_csv(VISITOR_EDUCATIONAL_ATTAINMENT_file)
with gzip.open(VISITOR_EDUCATIONAL_ATTAINMENT_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    df_VISITOR_EDUCATIONAL_ATTAINMENT.to_csv(gzipped_file, index=False,sep="|")

df_VISITOR_PSYCHOGRAPHIC_PROFILES = pd.read_csv(VISITOR_PSYCHOGRAPHIC_PROFILES_file)
with gzip.open(VISITOR_PSYCHOGRAPHIC_PROFILES_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    df_VISITOR_PSYCHOGRAPHIC_PROFILES.to_csv(gzipped_file, index=False,sep="|")

df_VISITOR_AGE_RANGE = pd.read_csv(VISITOR_AGE_RANGE_file,sep="|")
with gzip.open(VISITOR_AGE_RANGE_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    df_VISITOR_AGE_RANGE.to_csv(gzipped_file, index=False,sep="|")

ft_data.info()

locations = ft_data[['STORE_ID','NAME','STREET_ADDRESS','ZIP_CODE']]
locations = locations.drop_duplicates()
locations = locations.reset_index()
locations.info()
locations['STORE_ID'].nunique()

locations = pd.merge(locations,
                     poi_sod[['PLACEKEY','POSTAL_CODE','STREET_ADDRESS','LOCATION_NAME']],
                     left_on=['NAME','STREET_ADDRESS','ZIP_CODE'],
                     right_on=['LOCATION_NAME','STREET_ADDRESS','POSTAL_CODE'],
                     how='inner')
locations = locations.drop_duplicates()
locations.info()
locations['STORE_ID'].nunique()
locations['PLACEKEY'].nunique()

visits = ft_data[['index','STORE_ID','MONTH_STARTING','MONTH_ENDING','TOTAL_VISITS']]
visits.loc[:,'MONTH_STARTING'] = visits['MONTH_STARTING'].str[:10]
visits.loc[:,'MONTH_ENDING'] = visits['MONTH_ENDING'].str[:10]
visits.head()


locations_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_locations.csv'
visits_file = 'C:/Users/dratnadiwakara2/Downloads/temp_dewey/retail_store_visitor_visits.csv'

with gzip.open(locations_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    locations.to_csv(gzipped_file, index=False,sep="|")
    
with gzip.open(visits_file+'.gz', 'wt', encoding='utf-8') as gzipped_file:
    visits.to_csv(gzipped_file, index=False,sep="|")
